refactor: new hobbyist architecture

+++ PeruStar.API/Security/Domain/Repositories/IUserRepository.cs
@@ -6,7 +6,7 @@ public interface IUserRepository
 {
     Task<IEnumerable<User>> ListAsync();
     Task AddAsync(User user);
-    Task<User> FindByIdAsync(int id);
+    Task<User> FindByIdAsync(long id);
     Task<User> FindByEmailAsync(string email);
     bool ExistsByEmail(string email);
     void Update(User user);
